temp commit

+++ src/main/java/fr/univlr/info/AppointmentAPIV1/controller/DoctorController.java
@@ -1,6 +1,8 @@
 package fr.univlr.info.AppointmentAPIV1.controller;
 
+import fr.univlr.info.AppointmentAPIV1.model.Appointment;
 import fr.univlr.info.AppointmentAPIV1.model.Doctor;
+import fr.univlr.info.AppointmentAPIV1.store.AppointmentRepository;
 import fr.univlr.info.AppointmentAPIV1.store.DoctorRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.boot.web.server.LocalServerPort;
@@ -24,6 +26,8 @@ import java.net.URI;
 import java.util.Collection;
 import java.util.List;
 import java.util.NoSuchElementException;
+import org.springframework.web.bind.annotation.RequestParam;
+
 
 @RestController
 @RequestMapping(path = "/api")
@@ -40,27 +44,11 @@ public class DoctorController {
         return new ResponseEntity<>(docs, HttpStatus.OK);
     }
 
-    @PostMapping("/doctors")
-    ResponseEntity<Doctor> newDoctor(@Valid @RequestBody Doctor doc) {
-        // if(this.docRepository.findById(doc.getId()).isPresent()) {
-        // return new ResponseEntity<>(HttpStatus.CONFLICT);
-        // }
-        Doctor Doctor = docRepository.save(doc);
-        HttpHeaders headers = new HttpHeaders();
-        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
-                .path("/{id}")
-                .buildAndExpand(Doctor.getId())
-                .toUri();
-        // URI location = URI.create("http://localhost:8000/api/Doctors");
-        headers.set("Accept", "application/json");
-        headers.setLocation(location);
-        return new ResponseEntity<>(Doctor, headers, HttpStatus.CREATED);
-    }
-
-    @GetMapping("/doctors/{id}")
-    public ResponseEntity<Doctor> getById(@PathVariable Long id) {
+    @GetMapping("/doctors/{name}")
+    public ResponseEntity<Doctor> getById(@PathVariable String name) {
+        //Test d'existence
         try {
-            Doctor doc = docRepository.findById(id).get();
+            Doctor doc = docRepository.findByName(name).get();
             return new ResponseEntity<>(doc, HttpStatus.OK);
         } catch (NoSuchElementException e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -68,21 +56,22 @@ public class DoctorController {
 
     }
 
-    // @PutMapping("/doctors/{id}")
-    // public ResponseEntity<Doctor> updateDoctor(@PathVariable Long id, @Valid @RequestBody Doctor doc) {
-    //     Doctor olddoc = docRepository.findById(id).get();
-    //     olddoc.setDoctor(doc.getDoctor());
-    //     olddoc.setEndDate(doc.getEndDate());
-    //     olddoc.setPatient(doc.getPatient());
-    //     olddoc.setStartDate(doc.getStartDate());
-    //     Doctor updateddoc = docRepository.save(olddoc);
-    //     return new ResponseEntity<>(updateddoc, HttpStatus.OK);
-    // }
+    @GetMapping("/doctors/{name}/appointments")
+    public ResponseEntity<List<Appointment>> getAllDocAppointments(@PathVariable String name) {
+        Doctor doc = docRepository.findByName(name).get();
+        List<Appointment> appts = doc.getAppointments();
+        return new ResponseEntity<>(appts,HttpStatus.OK);
+    }
+    
 
-    @DeleteMapping("/doctors/{id}")
-    public ResponseEntity<Boolean> deleteDoctor(@PathVariable Long id) {
+    @DeleteMapping("/doctors/{name}")
+    public ResponseEntity<Boolean> deleteDoctor(@PathVariable String name) {
+        //Test d'existence
         try {
-            Doctor doc = docRepository.findById(id).get();
+            Doctor doc = docRepository.findByName(name).get();
+            //test d'intégrité
+            if(doc.getAppointments().size() != 0)
+                return new ResponseEntity<>(false,HttpStatus.CONFLICT);
             docRepository.delete(doc);
             return new ResponseEntity<>(true, HttpStatus.OK);
         } catch (NoSuchElementException e) {
@@ -92,7 +81,7 @@ public class DoctorController {
     }
 
     @DeleteMapping("/doctors")
-    public ResponseEntity<Boolean> deleteAllDoctor(@PathVariable(required = false) Long id) {
+    public ResponseEntity<Boolean> deleteAllDoctor() {
         docRepository.deleteAll();
         return new ResponseEntity<>(true, HttpStatus.OK);
     }
